+++ packages/components/package.json
@@ -62,6 +62,7 @@
         "@radix-ui/react-toggle-group": "^1.1.0",
         "@radix-ui/react-tooltip": "^1.1.2",
         "downshift": "^9.0.6",
+        "react-colorful": "^5.6.1",
         "tailwind-merge": "^2.4.0",
         "tailwind-variants": "^0.2.1"
     },
