+++ packages/components/src/components/ColorPicker/ColorPicker.tsx
@@ -7,22 +7,16 @@ import { SegmentedControl } from '../SegmentedControl/SegmentedControl';
 
 import { ColorPreview } from './ColorPreview';
 import styles from './styles/colorPicker.module.scss';
-import { type Color, type ColorPickerProps } from './types';
+import { type ColorPickerProps } from './types';
 
 type PickerType = 'brand' | 'custom';
 
-export const ColorPicker = ({
-    currentColor,
-    onColorChange,
-    showPreview = true,
-    allowCustomColor = true,
-    children,
-}: ColorPickerProps) => {
+export const ColorPicker = ({ showPreview = true, children, ...props }: ColorPickerProps) => {
     const [activePicker, setActivePicker] = useState<PickerType>('custom');
 
     return (
         <div className={styles.root}>
-            {showPreview && <ColorPreview color={currentColor} />}
+            {showPreview && <ColorPreview color={props.currentColor} />}
             <div className={styles.pickerSwitcher}>
                 <SegmentedControl.Root
                     defaultValue="custom"
@@ -39,17 +33,8 @@ export const ColorPicker = ({
                 </SegmentedControl.Root>
             </div>
             <div className={styles.pickers} data-active-picker={activePicker}>
-                {Children.map(children, (child, index) => (
-                    <RadixSlot
-                        key={index}
-                        data-test="tttt"
-                        currentColor={currentColor}
-                        onColorChange={(color: Color) => {
-                            onColorChange(color);
-                        }}
-                    >
-                        {child}
-                    </RadixSlot>
+                {Children.map(children, (child) => (
+                    <RadixSlot {...props}>{child}</RadixSlot>
                 ))}
             </div>
         </div>
